Support specifying a plugin module to perform additional checks.
Right now only three extension points exist: "call", "destructure",
and "symbol-to-expression", as these are the only ones needed to
support the two example linters. The first linter checks that any
reference to fields in a local which is directly bound to a module
from `require` happens on modules which actually exist on that
module. The second is a compile-time arity check which happens for any
module which has `arity-check?` set to true on its metatable.