Introduce "self-hosted" fennel!!!
This means that the outer "shell" of the Fennel module is written in
Fennel, and it loads all the real juicy implementation bits from
utils.lua, parser.lua, compiler.lua, and specials.lua, and the repl
source and macro source are all just strings still.

And the macros don't actually get loaded.

BUT STILL.